build(webpages): Build html webpages for recipes project with Odin
This is the start of a project I am building with the Odin Project where I am using my skills  I learned  in the foundations. I made 2 pages so far , and added links to some of my favorite recipes.  Note the picture don recipes.html will need to be fixed.