style(Alt Attribute): Fix alt attribute on all pictures to show correctly instead of txt on the browser
On each recipe picture, I fixed the alt attribute so that the description is hidden and not shown as text on the broswer. I treated each image as a  link before the changes. I added these alt attributes so that my pages would be more accessable to people who may be visually impaired or if  the image couldn't be displayed and people can still understand the context.