feat(Interactive Nav Change Menu): Change index navigation to change colour when hovering, experiment
Added a class on the ul (links-list) on the index to be referenced in the stylepage for a neat feature I saw  today that I wanted to try.  I made it so the links now change background colour when the mouse is hovered on them.  See the Feb11,2025 comments I made in the CSS page for the added in today.